--- a/source/dcv/features/corner/harris.d
+++ b/source/dcv/features/corner/harris.d
@@ -180,8 +180,7 @@ struct ShiTomasiDetector
 {
     @fastmath @nogc nothrow float opCall(float r1, float r2, float r3)
     {
-        import std.math : sqrt;
-
+        import ldc.intrinsics : sqrt = llvm_sqrt;
         return ((r1 + r3) - sqrt((r1 - r3) * (r1 - r3) + r2 * r2));
     }
 }