… source/dcv/imgproc/convolution.d
… source/dcv/imgproc/filter.d
… source/dcv/imgproc/imgmanip.d
… source/dcv/imgproc/interpolate.d
… source/dcv/imgproc/threshold.d
… source/dcv/multiview/stereo/matching.d
… tests/performance-tests/dub.json
… tests/performance-tests/source/performance/measure.d