This should be first optimization pass advised in #34 and #27, mostly by refactoring loops into mir.ndslice.algorithm iterations.

@9il I'll try making a commit per function (or related group of functions) when optimized. I'm making an early wip PR, so I can apply corrections from potential reviews  to further optimizations. Hope you're OK with this plan.

Progress:
- [x] dcv.features.corner.harris
- [x] dcv.features.utils.extractCorners
- [x] dcv.imgproc.color
- [x] dcv.imgproc.convolution
- [x] dcv.imgproc.filter.bilateralFilter
- [x] dcv.imgproc.filter.calcGradients
- [x] dcv.imgproc.filter.calcPartialDerivatives
- [x] dcv.imgproc.filter.canny
- [x] dcv.imgproc.filter.filterNonMaximum
- [x] dcv.imgproc.filter.nonMaximaSupression
- [x] dcv.imgproc.imgmanip.remap/warp
- [x] dcv.imgproc.threshold.threshold