@ljubobratovicrelja 2016-09-29
After d870aa60a3ad42eeb70ef17a0e50102f7c5c613e, running benchmark on my laptop yields:

```
$ ./benchmark.sh master -i 1000 -t harris

-----------------------------------------------------------------------------------
Profiling...
-----------------------------------------------------------------------------------

master:
dcv.features.corner.harris.harrisCorners.3:775 ms and 185 μs
dcv.features.corner.harris.harrisCorners.5:1 sec, 527 ms, and 822 μs
dcv.features.corner.harris.shiTomasiCorners.3:724 ms and 623 μs
dcv.features.corner.harris.shiTomasiCorners.5:1 sec, 529 ms, and 490 μs
master profiling done, results written in /home/relja/git/dcv/tests/performance-tests/.cache/master/tests/performance-tests/profile.csv

this:
dcv.features.corner.harris.harrisCorners.3:499 ms and 515 μs
dcv.features.corner.harris.harrisCorners.5:586 ms and 269 μs
dcv.features.corner.harris.shiTomasiCorners.3:431 ms and 233 μs
dcv.features.corner.harris.shiTomasiCorners.5:578 ms and 913 μs
 profiling done, results written in /home/relja/git/dcv/tests/performance-tests/profile.csv

-----------------------------------------------------------------------------------
Comparing and writing comparison results...
-----------------------------------------------------------------------------------

Comparison done, results written in /home/relja/git/dcv/tests/performance-tests/benchmark.csv

-----------------------------------------------------------------------------------
Results...
-----------------------------------------------------------------------------------

Function Name,Previous Runtime[μs],Current Runtime[μs],Speedup[%]
dcv.features.corner.harris.harrisCorners.3,775185,499515,55
dcv.features.corner.harris.harrisCorners.5,1527822,586269,160
dcv.features.corner.harris.shiTomasiCorners.3,724623,431233,68
dcv.features.corner.harris.shiTomasiCorners.5,1529490,578913,164
```