@9il 2016-09-29
Do you use mcpu=native flag for tests?
You may want to add fastmath