@ljubobratovicrelja 2016-09-29
> Are tests is single thread?

No - do you think they should be? I wanted to test raw performance, seemed like there's no need to run them in single thread. Anyhow this is easily done because of optional parallelization, so it might as well be defined by optional flag when running tests.

> Do you use mcpu=native flag for tests?

Oops, I was sure I added that... Not at the moment, will add it.