@ljubobratovicrelja 2016-09-29
> Good paper http://citeseerx.ist.psu.edu/viewdoc/download?doi=10.1.1.93.1608&rep=rep1&type=pdf

Will take a look, thanks.

I originally wanted to implement constant time median filtering described here:
https://nomis80.org/ctmf.pdf

AFAIK it is commonly best used with bigger kernels. It's also vectorizable like the one described in the paper you've posted, which is great. We'll definitely talk about this in future..