@9il 2016-09-30
> I originally wanted to implement constant time median filtering described here:
> https://nomis80.org/ctmf.pdf

Looks very good too, probably better for median filter