@ljubobratovicrelja 2016-10-01
Finished `dcv.imgproc.color` module. I was mainly extracting conversion algorithm code from the main function, to be easily replaced afterwards with std.experimental.color conversion implementation. And by that replacing loops and `lockstep` with `ndEach`, `ndMap` and `assumeSameStructure`. With `rgb2gray` and `gray2rgb` it'd be really nice to have [support for packed slices in assumeSafeStructure](https://github.com/ljubobratovicrelja/dcv/blob/0fc5b01bf29c4030541f253bcc3c0ed1f5edef49/source/dcv/imgproc/color.d#L88-L91). Overall improvement in speed is nice:

```
-----------------------------------------------------------------------------------
Results...
-----------------------------------------------------------------------------------

Function Name,Previous Runtime[μs],Current Runtime[μs],Speedup[%]
dcv.imgproc.color.gray2rgb,284986,259867,9
dcv.imgproc.color.hsv2rgb,8043678,1001873,702
dcv.imgproc.color.rgb2gray,3845346,489591,685
dcv.imgproc.color.rgb2hsv,6374634,1177531,441
dcv.imgproc.color.rgb2yuv,4144262,1067521,288
dcv.imgproc.color.yuv2rgb,3945694,950825,314
```