@9il 2016-10-01
First, the speed up number are awesome! I would be very happy to see your blog post about this optimizaiton! For example, at http://blog.mir.dlang.io/

> With rgb2gray and gray2rgb it'd be really nice to have support for packed slices in assumeSafeStructure.

The packed slices are not required here:
For example, for rgb2gray kernel algorithm will looks something like:
```d
...
{
    auto prealloc = uninitilizedSlice!(V[3])(slice.shape);
    assumeSameStructure!("a", "b")(prealloc, slice).ndEach!rgb2grayImpl();
    ...
    return cast(typeof(return)) prealloc;
}
void rgb2grayImpl(P)(P pointerTuple) // no need ref
{
    auto v = pointerTuple.b;
    pointerTuple.a[0] = v;
    pointerTuple.a[1] = v;
    pointerTuple.a[2] = v;
}
```