@9il 2016-10-01
Also, probably std.experimental.color can be used for some kernels