@ljubobratovicrelja 2016-10-02
> First, the speed up numbers are awesome! I would be very happy to see your blog post about this optimizaiton! For example, at http://blog.mir.dlang.io/

Yes - definitely. I think its very important for people to see how nd iterations are faster with ndslice.algorithm then with loops. This is important for Mir, but also very important for D - this example could mean a lot to newcomers searching for fast numerical code. I'll write it when we're finished with this PR, OK?

> For example, for rgb2gray kernel algorithm will looks something like:

Awesome, thanks!

> Also, probably std.experimental.color can be used for some kernels

Yes, for sure - I'll change this in following PRs.