@ljubobratovicrelja 2016-10-02
> assumeSameStructure requires identical strides, which is not true for sub-images.

Ah, tnx. Sorry didn't see that, ignore the comment above.