@ljubobratovicrelja 2016-11-11
@henrygouk could you take a look at the merge with master [diff](https://github.com/ljubobratovicrelja/dcv/pull/58/commits/824a4a893da35a3bec6ac9ae2f947ab7d4c8f7bb#diff-776b62583704f0aeefb3899150a16eaaR1165) (after bilateral sigma separation)? Selected line and those few below are the only edits, else is more-less taken from master. I believe I got the equation right, but I'd appreciate a double check. :)